pyproject.toml
[project]
name = "mcpserver-trail"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "langchain-mcp-adapters>=0.1.13",
    "langchain-openai>=1.0.3",
    "langgraph>=1.0.3",
    "python-dotenv>=1.2.1",
]
